+++ sh/build.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+cd code
+rm -rf ./lib
+node ../node_modules/typescript/bin/tsc